Unified execution base execution strategies https://github.com/SoftUni-Internal/interactive-issues/issues/250
