+++ OJS.Workers.Common/Extensions/CollectionExtensions.cs
@@ -1,0 +1,15 @@
+﻿namespace OJS.Workers.Common.Extensions
+{
+    using System.Collections.Generic;
+
+    public static class CollectionExtensions
+    {
+        public static void AddRange<T>(this ICollection<T> destination, IEnumerable<T> source)
+        {
+            foreach (var item in source)
+            {
+                destination.Add(item);
+            }
+        }
+    }
+}
